Se crearon permisos en el seedder y se les adiciono a los roles como de igual manera se les asigno los permisos a las urls de las vistas

+++ resources/views/administrador/usuarios/roles.blade.php
@@ -10,14 +10,16 @@
                         <div class="col">
                             <h4 class="card-title">
                                 <span class="badge bg-danger fs-4">
-                                     ROLES
+                                    ROLES
                                 </span>
                             </h4>
                         </div>
                         <div class="col-auto">
-                            <button class="btn btn-primary" onclick="abrirModalRol()">
-                                <i class="fas fa-plus me-1"></i> Nuevo
-                            </button>
+                            @can('admin.rol.crear')
+                                <button class="btn btn-primary" onclick="abrirModalRol()">
+                                    <i class="fas fa-plus me-1"></i> Nuevo
+                                </button>
+                            @endcan
                         </div>
                     </div>
                 </div>
@@ -45,21 +47,29 @@
                                 class="rounded-circle thumb-xxl">
                         </div>
                         <div class="mt-3">
-                            <a href="#" class="btn btn-sm btn-outline-warning px-2 d-inline-flex align-items-center"
-                                onclick="abrirModalRol('{{ $lis->id }}')">
-                                <i class="iconoir-warning-circle fs-14 me-1"></i>
-                                Editar
-                            </a>
-                            <a href="#" class="btn btn-sm btn-outline-danger px-2 d-inline-flex align-items-center"
-                                onclick="eliminarRol('{{ $lis->id }}')">
-                                <i class="iconoir-trash fs-14 me-1"></i>
-                                Eliminar
-                            </a>
+                            @can('admin.rol.editar')
+                                <a href="#" class="btn btn-sm btn-outline-warning px-2 d-inline-flex align-items-center"
+                                    onclick="abrirModalRol('{{ $lis->id }}')">
+                                    <i class="iconoir-warning-circle fs-14 me-1"></i>
+                                    Editar
+                                </a>
+                            @endcan
+
+                            @can('admin.rol.eliminar')
+                                <a href="#" class="btn btn-sm btn-outline-danger px-2 d-inline-flex align-items-center"
+                                    onclick="eliminarRol('{{ $lis->id }}')">
+                                    <i class="iconoir-trash fs-14 me-1"></i>
+                                    Eliminar
+                                </a>
+                            @endcan
+                            
+                            @can('admin.rol.visualizar')
                             <a href="#" class="btn btn-sm btn-outline-secondary px-2 d-inline-flex align-items-center"
                                 onclick="vizualizarRolPermiso('{{ $lis->id }}')">
                                 <i class="iconoir-eye fs-14 me-1"></i>
                                 Vizualizar
                             </a>
+                            @endcan
                         </div>
                     </div>
                 </div>
